% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epfac.R
\name{epfac}
\alias{epfac}
\title{Calculate epistasis factor}
\usage{
epfac(object, background = "P2", n = 1000)
}
\arguments{
\item{object}{fitted model object of the format linecross}

\item{background}{the genetic bakground in which the epistasis factor should be calculated}

\item{n}{number of samples in the Montecarlo simulation}
}
\value{
The epistasis factor with uncertainty
}
\description{
\code{epfac} calculates the epistasis factor of a genetic background given a model
}
\details{
The epistasis factor measures how a genetic background modifies the reference effects. The
uncertainty in the epistasis factor is estimated using Monte carlo simulation assuming
that the estimated genetic parameters have an uncertainty that is multivariate normal.
}
\examples{
## Run the model function linecross_models
mod=lcross(model="general", ref="P1", data=ipomopsis, maxeval=3)

## Calculate the epistasis factor
epfac(object = mod, background = "P2", n = 10)
# NB you should use n = 1000 or more to get an accurate error estimate

}
\author{
Geir H. Bolstad & Salomé Bourg
}
