% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipomopsis.R
\docType{data}
\name{ipomopsis}
\alias{ipomopsis}
\title{Campbell et al. (2008) data}
\format{
\subsection{}{

A data frame with 9 rows and 3 columns:
\describe{
\item{pop}{name of each experimental population}
\item{mean}{the poroporiton of survived seeds}
\item{se}{standard error of each mean}

}
}
}
\source{
\url{https://doi.org/10.1111/j.1558-5646.2008.00460.x}
}
\usage{
ipomopsis
}
\description{
The data is from a cross between two species of \emph{Ipomopsis}, the scarlet gilia \emph{I. aggregata}
and the slendertube skyrocket \emph{I. tenuituba}. The trait is proportion of survived seeds
surviving until reproduction planted in the environment native to \emph{I. aggregata}. The values
were obtained from Figure 3A in Campbell et al. (2008).
}
\references{
Campbell, D. R., Waser, N. M., Aldridge, G. & Wu, C. A. (2008) Lifetime fitness in
two generations of \emph{Ipomopsis} hybrids Evolution 62:2616-2627.
}
\keyword{datasets}
