% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcross.R
\name{lcross}
\alias{lcross}
\title{linecross model fitting}
\usage{
lcross(model, ref, data, maxeval = 3)
}
\arguments{
\item{model}{a model to fit is required. See ‘Details’. Currently, '"additive"',  '"dominance"',  '"add_dom"',
'"general"', '"general_dom"', '"generalWB"', '"generalWB_dom"', '"generalW"', '"generalW_dom"', '"generalB"',
'"generalB_dom"', '"classic"','"multilinear"', '"canalization"' and '"multilinear_add"'.}

\item{ref}{the reference population (can be P1, P2, F1 or F2) as a character}

\item{data}{data.frame with three columns (name of population, mean, standard error). See ipomopsis for an example.}

\item{maxeval}{for the non-linear models ('"multilinear"', '"canalization"' and '"multilinear_add"'), an optimization algorithm is used
to estimate the parameters requiring starting values. A grid search is performed on the values Yh, Y1 and Y2, their respective log
likelihood is computed for each point in the grid. If the grid-search has highest likelihood for extreme values of Yh, Y1 and/or Y2, the search
grid is expanded in this direction(s). The maximum number of new search grids produced is defined by 'maxeval'.}
}
\value{
fitted model object of the format linecross_models
}
\description{
\code{lcross} fits different models derived from the multilinear model of gene
interaction adapted to linecross data.
}
\details{
The data need to be in the required format for the function to run. See the tribolium vignette.

Model '"multilinear"' is useful for studying directional epistasis in the divergence of one population
from another. For this question it makes most sense to have the ancestral population as P1 or P2 and use
this as the reference. This model is non-linear.

Model '"multilinear_add"' is a modified version of the multilinear model without dominance effect. This
model is non-linear.

Model '"canalization"' is a modified version of the multilinear model can be used to study whether the
parental populations are canalized (i.e. have a flatter GP-map) compared to the hybrid populations. This
model makes most sense with F2 as the reference background in which to compare the parental populations.
This model is non-linear.

Models '"additive"', "dominance"', and the combined version'"add_dom" are simple models that respectively
only consider additive effect, dominance effect or both.

Model '"classic"' is the classic line cross model parameterized with additive (alpha) and dominance (delta)
effects rather than effects of substitutions (Lynch and Walsh 1998), where Eaa, Ead, and Edd are AxA, AxD and
DxD epsitasis components respectively.

Model '"general"' is the most general model in the family. When P1 is the reference, Ehh describes epistatic
interactions between heterozygote genotypes in the background of the P1 , Eh2 describes epistatic interactions
between P2 homozygotes and heterozygotes , and E22 describes epistatic interactions between P2 homozygotes.
This model is most useful as general model to study deviances from multilinearity.

Model '"general_dom"' is a modified version of the general epistasis model without additive effect.

Model '"generalWB"' is a modified version of the general epistasis model that specifically tests the
difference between within and between population epistasis in the background of the reference hybrid population.
This model makes most sense with an hybrid population (F1 or F2) as the reference background.

Model '"generalWB_dom"' is a modified version of the '"generalWB"' model without additive effect.

Model '"generalW"'is a modified version of the '"generalWB"' model that only estimates the within population
epistasis in the background of the reference population.

Model '"generalW_dom"' is a modified version of the '"generalW"' model without additive effect.

Model '"generalB"' is a modified version of the '"generalWB"' model that only estimates the between population
epistasis in the background of the reference population.

Model '"generalB_dom"' is a modified version of the '"generalB"' model without additive effect.
}
\examples{
lcross(model="general", ref="F1", data=ipomopsis, maxeval=3)

}
\references{
Bolstad, G. H., Bourg, S., Griffin, D., Pélabon, C. & Hansen, T. F. (2023) Quantifying
genome-wide functional epistasis by line-cross analysis. in prep.

Hansen, T. F., & Wagner, G. P. (2001)
Modeling genetic architecture: a multilinear theory of gene interaction.
Theoretical population biology. 59:61-86.

Lynch, M., & Walsh, B. (1998) Genetics and analysis of quantitative traits
(Vol. 1, pp. 535-557). Sunderland, MA: Sinauer.
}
\author{
Geir H. Bolstad & Salomé Bourg
}
