% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linecross.R
\name{plot_linecross}
\alias{plot_linecross}
\title{3D plot of line-cross GP-map}
\usage{
plot_linecross(output.model, data, theta, phi, margins)
}
\arguments{
\item{output.model}{fitted model object of the format linecross}

\item{data}{data.frame with three columns (derivatives, means, se)}

\item{theta}{same as function \code{\link{persp}}}

\item{phi}{same as function \code{\link{persp}}}

\item{margins}{same as function \code{mar} in \code{\link{par}}}
}
\value{
A plot of line-cross GP-maps
}
\description{
\code{linecross} uses multilinear model ...please write a proper description later
}
\examples{
#put an example there soon

}
\author{
Geir H. Bolstad & Salomé Bourg
}
