% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribolium.R
\docType{data}
\name{tribolium}
\alias{tribolium}
\title{Demuth, J. P., & Wade, M. J. (2007) data}
\format{
\subsection{}{

A data frame with 16 rows and 3 columns:
\describe{
\item{derivatives}{P1, P2, F1_12 , F1_21, F2_12 , F2_11 , F2_22 , F2_21 ,B1_12 , B1_11a, B1_11b, B1_21 ,
B2_21 , B2_22a, B2_22b, B2_12}
\item{means}{the number of offspring (at 29 degrees)}
\item{standard errors}{se from the means}

}
}
}
\source{
\url{https://doi.org/10.1111/j.1558-5646.2007.00048.x}
}
\usage{
tribolium
}
\description{
The data from a cross between two populations (Mexico x Tanzania) of the flour beetle Tribolium castaneum,
where the full spectrum of line crosses was realized (two parental populations P1 and P2, two reciprocal F1’s,
four F2’s, and eight backcrosses). They describe the genetic basis of divergence among populations of
Tribolium castaneum. The life-history trait studied is the number of offspring (at 29 degrees).
}
\references{
Demuth, J. P., & Wade, M. J. (2007).
Population differentiation in the beetle Tribolium castaneum. I. Genetic architecture.
Evolution, 61(3), 494-509.
}
\keyword{datasets}
