% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linecross.R
\name{plot_linecross}
\alias{plot_linecross}
\title{3D plot of line-cross GP-map}
\usage{
plot_linecross(
  object,
  theta = -40,
  phi = 20,
  margins = c(2, 2, 2, 2),
  col.triangle = "gray80",
  main = "default",
  cex.text = 1.2,
  ticktype = "simple",
  zlab = "z"
)
}
\arguments{
\item{object}{fitted model object of the format linecross}

\item{theta}{as in function \code{\link{persp}}}

\item{phi}{as in function \code{\link{persp}}}

\item{margins}{as in function \code{mar} in \code{\link{par}}}

\item{col.triangle}{colors for the lines of the triangle formed by linecross analysis. See details.}

\item{main}{as in function \code{\link{plot}}. The default value is the model name.}
}
\value{
A 3D plot of line-cross GP-map
}
\description{
\code{plot_linecross} draws a 3D plot of linecross GP-map. Black dots represent the observed mean for
the populations and derivatives and grey dots are the predicted values.
}
\details{
The axes are genotypes (proportion of P2 alleles, denoted S), heterozygosity (the probability that
an individual has one P1 and one P2 allele at a locus, denoted H) and phenotypes. The two variables
S and H define a two-dimensional genetic-content space, with P1, F1 and P2 at the edges. Using these
variables in a regression with the phenotype as the response variable, this triangle becomes a surface
where the height gives the genotypic value (i.e. the trait value of each genotype). This surface is a
description of a two-dimensional genotype-phenotype map representing the composite (genome wide sum)
genetic effects.

If the surface is flat and tilted in the S-direction this indicates additive effects. A flat surface
with tilt in the H-direction indicates dominance. Any non-linearities in the surface would indicate
epistasis, and different types of epistasis will generate different non-linear surfaces.
}
\examples{
## Run the model function linecross_models
mod=lcross(model="general", reference="F1", data=tribolium, maxeval=3)

## Plot the corresponding linecross GPmap
plot_linecross(object=mod, data=tribolium, theta=-60, phi=20,
margins=c(2,2,2,2), col.triangle="gray80")

}
\seealso{
for X-Y-Z plotting see ‘persp’
}
\author{
Geir H. Bolstad & Salomé Bourg
}
